requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.1+cpu 
torchvision==0.10.1+cpu
Pillow==7.2.0
albumentations==0.5.2
numpy==1.19.5
tqdm==4.48.0
scikit-learn==0.24.0
glob2
gunicorn
uvicorn==0.15.0
python-multipart